#Week5 Chatbot/API/requirements.txt
fastapi
uvicorn
requests==2.28.2
numpy==1.24.2
pandas==1.5.3
pythainlp==3.1.1
rank-bm25==0.2.2
simpletransformers
python-crfsuite
stop-words
torchvision